requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
langgraph
langchain-core
yt-dlp
faster-whisper
streamlit
requests
pytest
httpx
